fix(SearchBar-ios): Stop layoutAnimation effects (#1979)
Changes in https://github.com/react-native-training/react-native-elements/commit/78ff4d81ee4b71a312a16a775b9438971a9280a8 mean that sometimes the JS thread will update the UI based on the onChangeText callback and do the layout animation for the SearchBar together.

This results in any content that depends on the onChangeText callback getting this animation applied. This fix allows the onChangeText callback and the blur to run in different frames.